Add files via upload
A program used to calculate the number of hours in the input(num_of_days). The program also validates the input logic to make sure it is within the rules of a positive number and not any other valid than a valid day input. The program will print the number of hours and the days entered or will print invalid if the answer is not a relevant value.